Seperated WorldSoundService out of the WorldSoundPlayer
